HHH-14503 - Migrate tests from jpa/test to orm/test/jpa
